+++ src/Symfony/Component/Form/Extension/Validator/Type/FieldTypeValidatorExtension.php
@@ -13,9 +13,12 @@ namespace Symfony\Component\Form\Extension\Validator\Type;
 
 use Symfony\Component\Form\AbstractTypeExtension;
 use Symfony\Component\Form\FormBuilder;
-use Symfony\Component\Form\Extension\Validator\Validator\DelegatingValidator;
+use Symfony\Component\Form\Extension\Validator\EventListener\DelegatingValidationListener;
 use Symfony\Component\Validator\ValidatorInterface;
 
+/**
+ * @author Bernhard Schussek <bschussek@gmail.com>
+ */
 class FieldTypeValidatorExtension extends AbstractTypeExtension
 {
     private $validator;
@@ -39,7 +42,8 @@ class FieldTypeValidatorExtension extends AbstractTypeExtension
             ->setAttribute('validation_groups', $options['validation_groups'])
             ->setAttribute('validation_constraint', $options['validation_constraint'])
             ->setAttribute('cascade_validation', $options['cascade_validation'])
-            ->addValidator(new DelegatingValidator($this->validator));
+            ->addEventSubscriber(new DelegatingValidationListener($this->validator))
+        ;
     }
 
     public function getDefaultOptions()
